Expense_tracking_system/requirements.txt
datetime
numpy
pandas
openpyxl
matplotlib
seaborn
requests
mysql-connector-python
pytest
pydantic
streamlit